% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dtest_compute_D.R
\name{Dtest_compute_D}
\alias{Dtest_compute_D}
\title{Function to read the posterior predictive samples and compute the D test values.}
\usage{
Dtest_compute_D(
  t1,
  t2,
  tag = "test",
  dir_files = "./partials",
  only_reps = NULL
)
}
\arguments{
\item{t1}{a list of stochastic maps.}

\item{t2}{a list of stochastic maps.}

\item{tag}{a name (pattern) used to name the files produced by the simulation.}

\item{dir_files}{the directory where the files will be save. You WILL need to know where this location is!}

\item{only_reps}{a numeric vector with the indexes of the replicates to read from the directory. If NULL, then the function will expect 'length(t1)' replicates (length(t1) == length(t2) ). Use this parameter to control which replicates will be used to compute the D test statistics.}
}
\value{
a list with the D test results.
}
\description{
Note that Both T1 and T2 replicates need to be in the same directory. Note that the final test result will have sample size equal to the number of replicates in the directory. This function will not perform simulations.
}
