% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dtest_MCMC_informed_root.R
\name{Dtest_MCMC_informed_root}
\alias{Dtest_MCMC_informed_root}
\title{Function to perform the MCMC posterior predictive simulations}
\usage{
Dtest_MCMC_informed_root(
  t1,
  t2,
  tag = "test",
  dir_files = "./partials",
  ngen = 5000,
  nsim = 1000,
  from_to = NULL,
  read_only = FALSE,
  model_mat1,
  model_mat2,
  root_p1,
  root_p2,
  prior_exp = 0.5,
  scale_mcmc = 1
)
}
\arguments{
\item{t1}{a list of stochastic maps.}

\item{t2}{a list of stochastic maps.}

\item{tag}{a name (pattern) used to name the files produced by the simulation.}

\item{dir_files}{the directory where the files will be save. You WILL need to know where this location is!}

\item{ngen}{number of generations for each MCMC chain.}

\item{nsim}{number of posterior predictive simulation replicates.}

\item{from_to}{index of the posterior predictive replicates to run. Can be used to structure an analysis over multiple processors and/or computers.}

\item{read_only}{if the function should only read the existing files instead of producing the simulations.}

\item{model_mat1}{the model for the trait 1. Same format as in the "model" parameter for the function "ace" in the package "ace".}

\item{model_mat2}{the model for the trait 2. Same format as in the "model" parameter for the function "ace" in the package "ace".}

\item{root_p1}{the vector of root probabilities for trait 1.}

\item{root_p2}{the vector of root probabilities for trait 2.}

\item{prior_exp}{the parameter for the exponential prior for the rates of transition of the Markov model fitted to the data.}

\item{scale_mcmc}{the scaling factor for the Metropolis step of the MCMC. See function "metrop" in package "mcmc".}
}
\value{
write a large number of ".rds" files to the directory. Please make sure that you have read and write permissions to the "dir_files" location! Also make sure to record the location of the "dir_files" directory.
}
\description{
Function to perform the MCMC posterior predictive simulations
}
