% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dtest_pred_sims.R
\name{Dtest_pred_sims}
\alias{Dtest_pred_sims}
\title{Function to perform the MCMC posterior predictive simulations}
\usage{
Dtest_pred_sims(
  t1,
  t2,
  tag = "test",
  dir_files,
  ngen = 5000,
  nsim = 100,
  only_reps = NULL,
  model_mat1,
  model_mat2,
  root_p1,
  root_p2,
  prior_exp = 0.5,
  scale_mcmc = 1
)
}
\arguments{
\item{t1}{a list of stochastic maps. Length of t1 and t2 need to be the same.}

\item{t2}{a list of stochastic maps. Length of t1 and t2 need to be the same.}

\item{tag}{a name (pattern) used to name the files produced by the simulation.}

\item{dir_files}{the directory where the files will be save. You WILL need to know where this location is!}

\item{ngen}{number of generations for each MCMC chain.}

\item{nsim}{number of stochastic mapping samples from each replicate.}

\item{only_reps}{a numeric vector with the indexes of the replicates to run. If NULL, then the function will do a number of replicates equal to the length of t1 ( length(t1) == length(t2) ). Use this parameter to spread replicates through multiple processors or computers.}

\item{model_mat1}{the model for the trait 1. Same format as in the "model" parameter for the function "ace" in the package "ace".}

\item{model_mat2}{the model for the trait 2. Same format as in the "model" parameter for the function "ace" in the package "ace".}

\item{root_p1}{the vector of root probabilities for trait 1.}

\item{root_p2}{the vector of root probabilities for trait 2.}

\item{prior_exp}{the parameter for the exponential prior for the rates of transition of the Markov model fitted to the data.}

\item{scale_mcmc}{the scaling factor for the Metropolis step of the MCMC. See function "metrop" in package "mcmc".}
}
\value{
Function will write a large number of ".rds" files to the directory. Please make sure that you have read and write permissions to the "dir_files" location! Also make sure to record the location of the "dir_files" directory. It also returns a list with information on replicates with problems.
}
\description{
This function will only run the simulations. After running the posterior predictive simulations, another function will be used to compute the test statistics. Note that the function will use rejection sampling to make sure that each predictive simulation has the same number of states as the original data (maybe we can relax this assumption and add a parameter to control this).
}
